refactor(services): separate container lifecycle responsibilities
