+++ rust-srec/docs/zh/release-notes/unreleased.md
@@ -2,6 +2,12 @@
 
 ## `unreleased`
 
+## 服务生命周期组织
+
+- **启动耗时记录聚焦初始化工作**
+
+  容器的关闭期限、输出根目录辅助逻辑、事件决策和现有生命周期测试已按职责拆分。启动日志不再记录细小的同步构造耗时，保留 I/O、引擎发现和整体耗时。初始化顺序、共享输出根目录快照及关闭行为保持不变；请参阅[容器职责](../concepts/architecture.md#服务容器职责)。
+
 ## 后端模型清理
 
 - **统一主播状态类型，移除未使用的 Rust 接口**
